refactor: rename telemetry namespace to instrumentation
